+++ bracketcolors/src/BracketMap.cc
@@ -1,46 +1,29 @@
-/* -----------------------------------------------------------------------------
+/*
+ *      BracketMap.cc
+ *
+ *      Copyright 2013 Asif Amin <asifamin(at)utexas(dot)edu>
+ *
+ *      This program is free software; you can redistribute it and/or modify
+ *      it under the terms of the GNU General Public License as published by
+ *      the Free Software Foundation; either version 2 of the License, or
+ *      (at your option) any later version.
+ *
+ *      This program is distributed in the hope that it will be useful,
+ *      but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *      MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *      GNU General Public License for more details.
+ *      You should have received a copy of the GNU General Public License
+ *      along with this program; if not, write to the Free Software
+ *      Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
+ *      MA 02110-1301, USA.
+ */
 
-    Asif Amin
-    asifamin@utexas.edu
-
-    This program is free software; you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation; either version 2 of the License, or
-    (at your option) any later version.
-
-    This program is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-    You should have received a copy of the GNU General Public License
-    along with this program; if not, write to the Free Software
-    Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
-    MA 02110-1301, USA.
-
------------------------------------------------------------------------------ */
-
-
-/* --------------------------------- INCLUDES ------------------------------- */
-/* ----- Platform ------- */
-#include <glib.h>
-/* ----- Standard ------- */
 #include <stack>
 #include <set>
-/* ------- User --------- */
+
+#include <glib.h>
+
 #include "BracketMap.h"
-
-/* ----------------------------------- USES --------------------------------- */
-
-/* ----------------------------------- TYPES -------------------------------- */
-
-/* --------------------------------- CONSTANTS ------------------------------ */
-
-/* ---------------------------------- EXTERNS ------------------------------- */
-
-/* --------------------------------- PROTOTYPES ----------------------------- */
-
-/* ------------------------------ IMPLEMENTATION ---------------------------- */
-
 
 
 // -----------------------------------------------------------------------------
