+++ bracketcolors/src/Makefile.am
@@ -6,8 +6,7 @@ geanyplugins_LTLIBRARIES = bracketcolors.la
 bracketcolors_srcs = \
     bracketcolors.cc \
     BracketMap.cc \
-    BracketMap.h \
-    utils.h
+    BracketMap.h
 
 bracketcolors_la_SOURCES = $(bracketcolors_srcs)
 bracketcolors_la_CXXFLAGS = $(GEANY_CFLAGS) $(AM_CXXFLAGS) $(AM_CPPFLAGS) -DG_LOG_DOMAIN=\"BracketColors\"
